functionary/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via jsonschema
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements.in
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
deprecated==1.2.13
    # via redis
django==4.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-spectacular
docker==6.0.0
    # via -r requirements.in
drf-spectacular[sidecar]==0.23.1
    # via -r requirements.in
drf-spectacular-sidecar==2022.8.1
    # via drf-spectacular
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
kombu==5.2.4
    # via celery
packaging==21.3
    # via
    #   docker
    #   redis
prompt-toolkit==3.0.30
    # via click-repl
pydantic==1.9.2
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytz==2022.2.1
    # via
    #   celery
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.in
    #   drf-spectacular
redis==4.3.4
    # via -r requirements.in
requests==2.28.1
    # via docker
six==1.16.0
    # via click-repl
sqlparse==0.4.2
    # via django
typing-extensions==4.3.0
    # via pydantic
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.12
    # via
    #   docker
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
websocket-client==1.3.3
    # via docker
wrapt==1.14.1
    # via deprecated
